Convert each handler function to 1 file
